refactor(hutch): split the subscription DynamoDB provider into read and write halves
`initDynamoDbSubscriptionProviders` mixed the read lookups (findByUserId,
findBySubscriptionId) and the six write mutations behind one factory. The save
gate only needs the reads, and the reads must stay in hutch (the gate computes
write access locally, never over HTTP), while the writes are destined to move
to the subscription service.

Extract `initDynamoDbSubscriptionRead` and `initDynamoDbSubscriptionWrites`
over a shared `subscription-provider-row` module (the Zod row schema +
`toRecord`). The combined `initDynamoDbSubscriptionProviders` becomes a thin
composition of the two, kept for the Lambda/CLI entry points that consume the
whole surface. The web composition root now assembles the two halves directly,
so a later swap of the write half for an HTTP adapter leaves the read half
untouched.

`deps.subscriptionProviders` keeps the same eight-method shape, so every
consumer and the per-op provider tests are unchanged.
